fix to enable multi-page routing
Cloudfront does not append the `.html` extension when it makes the
request to S3. Pages in S3 have the `.html` extension by default. This
causes a 404 error if you go directly to any page that's not the root.
This fix removes the `.html` extension from NextJS output so that it
matches Cloudfront's request.